+++ mySQLPunk/template/IConnectionDraftForm.cs
@@ -1,0 +1,9 @@
+using System.Collections.Generic;
+
+namespace mySQLPunk.template
+{
+    public interface IConnectionDraftForm
+    {
+        void ApplyConnectionDraft(Dictionary<string, object> connection);
+    }
+}